+++ mantra/datasets/__init__.py
@@ -1,4 +1,5 @@
 from .mantra import ManifoldTriangulations
 from .mantra_dataset import MantraDataset
+from .pachner_walks import PachnerWalkDataset
 
-__all__ = ["ManifoldTriangulations", "MantraDataset"]
+__all__ = ["ManifoldTriangulations", "MantraDataset", "PachnerWalkDataset"]